ffmuc-ebpf-clat: Update eBPF code
Now updates checksums and is designed to be attached to the WireGuard tunnel.

The idea is that the Linux network stack handles and forwards the packet as
IPv4 (which makes it easier to reuse existing firewalls, DHCP logic etc.),
and just before it enters the WG tunnel, we use the egress hook to rewrite
it to use the NAT64 prefix.

Return traffic hits the ingress hook and is rewritten to IPv4, so that the
kernel once again sees IPv4 and can forward it normally.

This way, all the NAT64 magic is kept close to the WireGuard side (which
already is Parker magic), so that the impact on the remaining "classic" Gluon
network stack is hopefully minimized.